Require ownership of photos

+++ app/routes/photo_routes.js
@@ -2,18 +2,21 @@
 
 // Express docs: http://expressjs.com/en/api.html
 const express = require('express')
+const passport = require('passport')
 const multer = require('multer') // set up multer for file photo
 const photo = multer({ dest: 'photos/' })
 const s3Create = require('../../lib/s3Create')
 const Photo = require('../models/photo') // pull in Mongoose model for photos
 const customErrors = require('../../lib/custom_errors')
 const handle404 = customErrors.handle404 // for non-existant requests
+const requireOwnership = customErrors.requireOwnership
 // removes blank fields from `req.body`
 const removeBlanks = require('../../lib/remove_blank_fields')
+const requireToken = passport.authenticate('bearer', { session: false })
 const router = express.Router() // mini app that only handles routes
 
 // GET (index) request on /photos
-router.get('/photos', (req, res, next) => {
+router.get('/photos', requireToken, (req, res, next) => {
   Photo.find()
     .then(photos => { // `photos` is an array of Mongoose documents
       return photos.map(photo => photo.toObject()) // converts each into a POJO
@@ -23,7 +26,7 @@ router.get('/photos', (req, res, next) => {
 })
 
 // GET (show) request on /photos/:id
-router.get('/photos/:id', (req, res, next) => {
+router.get('/photos/:id', requireToken, (req, res, next) => {
   Photo.findById(req.params.id) // req.params.id set based on `:id` in route
     .then(handle404)
     .then(photo => res.status(200).json({ photo: photo.toObject() }))
@@ -31,14 +34,14 @@ router.get('/photos/:id', (req, res, next) => {
 })
 
 // POST (create) request on /photos
-router.post('/photos', photo.single('attachment'), (req, res, next) => {
+router.post('/photos', photo.single('attachment'), requireToken, (req, res, next) => {
+  console.log('body', req.body)
+  console.log('user', req.user)
   const path = req.file.path
   const mimetype = req.file.mimetype
 
   s3Create(path, mimetype)
     .then(data => {
-      console.log('data: ', data)
-      // console.log('body: ', req.body)
       const url = data.Location
       const name = req.body.name
       let tags
@@ -47,11 +50,12 @@ router.post('/photos', photo.single('attachment'), (req, res, next) => {
       } else {
         tags = []
       }
-      // const owner =
+      const owner = req.user._id
       return Photo.create({
         name: name,
         url: url,
-        tags: tags
+        tags: tags,
+        owner: owner
       })
     })
     .then(photo => res.status(201).json({ photo: photo.toObject() }))
@@ -59,10 +63,12 @@ router.post('/photos', photo.single('attachment'), (req, res, next) => {
 })
 
 // PATCH (update) request on /photos/:id
-router.patch('/photos/:id', removeBlanks, (req, res, next) => {
+router.patch('/photos/:id', removeBlanks, requireToken, (req, res, next) => {
+  delete req.body.photo.owner
   Photo.findById(req.params.id)
     .then(handle404)
     .then(photo => {
+      requireOwnership(req, photo)
       return photo.updateOne(req.body.photo)
     })
     .then(() => res.sendStatus(204)) // on success return 204 and no JSON
@@ -70,11 +76,12 @@ router.patch('/photos/:id', removeBlanks, (req, res, next) => {
 })
 
 // DELETE (destroy) request on /photos/:id
-router.delete('/photos/:id', (req, res, next) => {
+router.delete('/photos/:id', requireToken, (req, res, next) => {
   Photo.findById(req.params.id)
     .then(handle404)
     .then(photo => {
-      // requireOwnership(req, photo) // not owner? throw error
+      requireOwnership(req, photo)
+      // not owner? throw error
       photo.deleteOne() // delete the photo ONLY IF the above didn't throw
     })
     .then(() => res.sendStatus(204)) // on success return 204 and no JSON
